+++ plugins/flashrom/meson.build
@@ -10,7 +10,6 @@ shared_module('fu_plugin_flashrom',
   fu_hash,
   sources : [
     'fu-flashrom-device.c',
-    'fu-flashrom-internal-device.c',
     'fu-plugin-flashrom.c',
     'fu-flashrom-cmos.c',
   ],
